Rename files based on feedback
