+++ omnisharptest/omnisharpIntegrationTests/advisor.integration.test.ts
@@ -8,7 +8,7 @@ import * as vscode from 'vscode';
 import { expect, should } from 'chai';
 import * as path from 'path';
 import { activateCSharpExtension, isRazorWorkspace, isSlnWithGenerator } from './integrationHelpers';
-import testAssetWorkspace from './testAssets/omnisharpTestAssetWorkspace';
+import testAssetWorkspace from './testAssets/activeTestAssetWorkspace';
 
 import { Advisor } from '../../src/features/diagnosticsProvider';
