+++ omnisharptest/omnisharpIntegrationTests/diagnostics.integration.test.ts
@@ -13,7 +13,7 @@ import {
     isSlnWithGenerator,
     restartOmniSharpServer,
 } from './integrationHelpers';
-import testAssetWorkspace from './testAssets/omnisharpTestAssetWorkspace';
+import testAssetWorkspace from './testAssets/activeTestAssetWorkspace';
 import { poll, assertWithPoll, pollDoesNotHappen } from './poll';
 import { isNotNull } from '../testUtil';
