+++ omnisharptest/omnisharpIntegrationTests/launchConfiguration.integration.test.ts
@@ -8,7 +8,7 @@ import * as vscode from 'vscode';
 
 import { should, expect } from 'chai';
 import { activateCSharpExtension, isRazorWorkspace, isSlnWithGenerator } from './integrationHelpers';
-import testAssetWorkspace from './testAssets/omnisharpTestAssetWorkspace';
+import testAssetWorkspace from './testAssets/activeTestAssetWorkspace';
 import { poll } from './poll';
 import { isNotNull } from '../testUtil';
